+++ src/core/include/vocotype/core/config.hpp
@@ -26,6 +26,7 @@ struct NormalizationConfig {
   bool compact_distances = true;
   bool currency_symbols = true;
   std::string punctuation_style = "chinese";
+  bool space_between_cjk_and_ascii = false;
 };
 
 struct OfflineAsrConfig {